Updates for new files for Travis-CI

+++ ssc/cmod_windpower.cpp
@@ -148,7 +148,7 @@ winddata::winddata(var_data *data_table)
 
 	double* rh = get_vector(data_table, "rh", &len);
 	if (rh != 0 && len == data.nrows() )
-		m_relativeHumidity = std::vector<float>(rh, rh+(int)len);
+		m_relativeHumidity = std::vector<double>(rh, rh+(int)len);
 	else m_relativeHumidity.clear();
 }
